+++ package.json
@@ -47,6 +47,7 @@
     "eslint-plugin-promise": "^4.2.1",
     "eslint-plugin-react": "^7.9.1",
     "eslint-plugin-standard": "^4.0.1",
+    "file-loader": "^4.2.0",
     "mini-css-extract-plugin": "^0.4.0",
     "postcss": "^7.0.17",
     "postcss-cli": "^6.1.3",
