+++ src/components/Cell/Readme.md
@@ -1,4 +1,4 @@
-```jsx
+```jsx harmony
   class Example extends React.Component {
 
     constructor() {
@@ -14,48 +14,47 @@
         <View activePanel={this.state.activePanel}>
           <Panel id="list">
             <PanelHeader>
-              Ячейки
+              Cell
             </PanelHeader>
-
-            <Group title="Простейший пример">
+            <Group header={<Header level="secondary">Простейший пример</Header>}>
               <List>
                 <Cell>Пятница</Cell>
                 <Cell>Суббота</Cell>
                 <Cell>Воскресение</Cell>
               </List>
             </Group>
-
-            <Group title="Переходы">
+            <Separator />
+            <Group header={<Header level="secondary">Переходы</Header>}>
               <List>
                 <Cell expandable onClick={() => this.setState({ activePanel: 'nothing' })}>Учетная запись</Cell>
                 <Cell expandable onClick={() => this.setState({ activePanel: 'nothing' })}>Основные</Cell>
                 <Cell expandable onClick={() => this.setState({ activePanel: 'nothing' })}>Приватность</Cell>
               </List>
             </Group>
-
-            <Group title="Индикатор">
+            <Separator />
+            <Group header={<Header level="secondary">Индикатор</Header>}>
               <List>
                 <Cell expandable onClick={() => this.setState({ activePanel: 'nothing' })} indicator="При использовании">Геолокация</Cell>
                 <Cell expandable onClick={() => this.setState({ activePanel: 'nothing' })} indicator="Всегда">Автопроигрывание медиа</Cell>
                 <Cell expandable onClick={() => this.setState({ activePanel: 'nothing' })} indicator="Выключены">Стикеры</Cell>
               </List>
             </Group>
-
-            <Group title="Многострочность">
+            <Separator />
+            <Group header={<Header level="secondary">Многострочность</Header>}>
               <List>
                 <Cell multiline>A Series of Unfortunate Events, Archer, Brooklyn Nine-Nine, Doctor Who, Game of Thrones</Cell>
                 <Cell multiline>The Avalanches</Cell>
               </List>
             </Group>
-
-            <Group title="Подпись">
+            <Separator />
+            <Group header={<Header level="secondary">Подпись</Header>}>
               <List>
                 <Cell description="Depeche Mode">Where’s the Revolution</Cell>
                 <Cell description="The Weeknd">I Feel It Coming (Feat. Daft Punk)</Cell>
               </List>
             </Group>
-
-            <Group title="Большая ячейка">
+            <Separator />
+            <Group header={<Header level="secondary">Большая ячейка</Header>}>
               <List>
                 <Cell
                   before={<Avatar size={72} />}
@@ -98,24 +97,24 @@
                   Екатерина Скобейко</Cell>
               </List>
             </Group>
-
-            <Group title="Иконки">
+            <Separator />
+            <Group header={<Header level="secondary">Иконки</Header>}>
               <List>
                 <Cell before={<Icon24About />}>Информация</Cell>
                 <Cell before={<Icon24Services />}>Сервисы</Cell>
               </List>
             </Group>
-
-            <Group title="Чекбоксы">
+            <Separator />
+            <Group header={<Header level="secondary">Чекбоксы</Header>}>
               <List>
                 <Cell selectable before={ <Avatar size={40} src="https://pp.userapi.com/c841034/v841034569/3b8c1/pt3sOw_qhfg.jpg" /> }>Артур Стамбульцян</Cell>
                 <Cell selectable before={ <Avatar size={40} src="https://pp.userapi.com/c841034/v841034569/3b8c1/pt3sOw_qhfg.jpg" /> }>Тимофей Чаптыков</Cell>
                 <Cell selectable before={ <Avatar size={40} src="https://pp.userapi.com/c841034/v841034569/3b8c1/pt3sOw_qhfg.jpg" /> }>Влад Анесов</Cell>
               </List>
             </Group>
-
+            <Separator />
             {this.state.removeList.length > 0 &&
-              <Group title="Удаление">
+              <Group header={<Header level="secondary">Удаление</Header>}>
                 <List>
                   {this.state.removeList.map((item, index) => (
                     <Cell key={item} removable onRemove={() => {
@@ -127,9 +126,9 @@
                 </List>
               </Group>
             }
-            
+            <Separator />
             {this.state.removeList.length > 0 &&
-              <Group title="Перетаскивание">
+              <Group header={<Header level="secondary">Перетаскивание</Header>}>
                 <List>
                   {this.state.draggingList.map((item) => (
                     <Cell key={item} draggable onDragFinish={({ from, to }) => {
