+++ src/components/CellButton/Readme.md
@@ -1,6 +1,6 @@
-```
+```jsx
 <View activePanel="button" header={false}>
-  <Panel id="button">
+  <Panel id="button" separator={false}>
     <Group title="Кнопка-ячейка">
       <CellButton>Добавить новую школу</CellButton>
     </Group>
