+++ src/components/Div/Readme.md
@@ -7,8 +7,9 @@
     <PanelHeader>Div</PanelHeader>
     <Group>
       <Div>
-        Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et
-        dolore magna aliqua.
+        <Button stretched level="secondary" size="l">
+          Edit Info
+        </Button>
       </Div>
     </Group>
   </Panel>
