+++ src/components/Group/Group.css
@@ -1,84 +1,25 @@
 .Group {
-  background: var(--background_content);
+  margin-bottom: 12px;
   }
 
   .Group__content {
     color: var(--text_primary);
     }
 
-/**
- * iOS
- */
-.Group--ios {
-  margin: 10px 0;
-  position: relative;
-  }
+  .Group__content:first-child {
+    margin-top: 12px;
+    }
 
-  .Group--ios .Group__content .List::before,
-  .Group--ios .Group__content .List::after {
-    content: none;
+  .Group__description {
+    font-size: 14px;
+    line-height: 16px;
+    color: var(--text_secondary);
     }
 
   .Group--ios .Group__description {
-    font-size: 14px;
-    line-height: 16px;
-    color: var(--text_secondary);
-    padding: 8px 12px;
+    padding: 4px 12px 24px;
     }
-
-  .Group--ios::before,
-  .Group--ios::after {
-    position: absolute;
-    right: 0;
-    left: 0;
-    z-index: 1;
-    height: 1px;
-    content: '';
-    background: var(--separator_common);
-    }
-
-  .Group--ios::before {
-    bottom: 100%;
-    transform-origin: center bottom;
-    }
-
-  .Group--ios::after {
-    top: 100%;
-    transform-origin: center top;
-    }
-
-  @media (min-resolution: 2dppx) {
-    .Group--ios::before,
-    .Group--ios::after {
-      transform: scaleY(.5);
-      }
-    }
-
-  @media (min-resolution: 3dppx) {
-    .Group--ios::before,
-    .Group--ios::after {
-      transform: scaleY(.33);
-      }
-    }
-
-/**
- * Android
- */
-
-.Group--android {
-  margin: 8px 0;
-  box-shadow: 0 2px 2px 0 rgba(0, 0, 0, .06), 0 0 2px 0 rgba(0, 0, 0, .03);
-  }
 
   .Group--android .Group__description {
-    font-size: 14px;
-    line-height: 16px;
-    color: var(--text_secondary);
-    padding: 12px 16px 8px;
+    padding: 4px 16px 24px;
     }
-
-.Panel__in-before + .Group--android,
-.PullToRefresh__content > .Group--android:first-child {
-  margin-top: 0;
-  padding-top: 8px;
-  }
