+++ src/components/Group/Group.js
@@ -2,16 +2,15 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import getClassName from '../../helpers/getClassName';
 import classNames from '../../lib/classNames';
-import Header from '../Header/Header';
 
 const baseClassNames = getClassName('Group');
 
 const Group = props => {
-  const { title, description, className, children, getRootRef, ...restProps } = props;
+  const { header, description, className, children, getRootRef, ...restProps } = props;
 
   return (
     <div {...restProps} ref={getRootRef} className={classNames(baseClassNames, className)}>
-      {title && <Header level="secondary">{title}</Header>}
+      {header}
       {children && <div className="Group__content">{children}</div>}
       {description && <div className="Group__description">{description}</div>}
     </div>
@@ -20,7 +19,7 @@ const Group = props => {
 
 Group.propTypes = {
   style: PropTypes.object,
-  title: PropTypes.node,
+  header: PropTypes.node,
   description: PropTypes.node,
   children: PropTypes.node,
   className: PropTypes.string,
@@ -28,11 +27,6 @@ Group.propTypes = {
     PropTypes.func,
     PropTypes.shape({ current: PropTypes.any })
   ])
-};
-
-Group.defaultProps = {
-  title: null,
-  description: null
 };
 
 export default Group;
