+++ src/components/HeaderContext/Readme.md
@@ -1,4 +1,4 @@
-```
+```jsx
 class Example extends React.Component {
 
   constructor () {
