+++ src/components/InfoRow/Readme.md
@@ -1,7 +1,7 @@
 Информационный блок. Используется для отрисовки некликабельных блоков с неизменяемым контентом (то есть без инпутов, слайдеров и т.п.)
 Если таких блоков несколько и они объединены по смыслу, то рекомендуется оборачивать список в Group с указанием title.
 
-```
+```jsx harmony
   <View activePanel="info-row">
     <Panel id="info-row">
       <PanelHeader>
@@ -14,6 +14,7 @@
           </InfoRow>
         </Div>
       </Group>
+      <Separator />
       <Group title="Информация о пользователе">
         <List>
           <Cell>
