+++ src/components/Link/Link.css
@@ -2,3 +2,13 @@
   color: var(--accent);
   text-decoration: none;
   }
+
+  .Link .Icon {
+    display: inline-block;
+    margin-left: 4px;
+    }
+
+  .Link .Icon--dropdown_16 {
+    position: relative;
+    top: 1px;
+    }
