+++ src/components/List/List.css
@@ -1,56 +1,1 @@
-.List {
-  position: relative;
-  margin: 0;
-  padding: 0;
-  background: var(--background_content);
-  }
-
-  .List .Cell,
-  .List .Radio {
-    position: relative;
-    z-index: 2;
-    }
-
-/* iOS */
-.List--ios {}
-
-  .List--ios::before,
-  .List--ios::after {
-    position: absolute;
-    right: 0;
-    left: 0;
-    z-index: 1;
-    height: 1px;
-    content: '';
-    }
-
-  .List--ios::before {
-    bottom: 100%;
-    background: var(--separator_common);
-    transform-origin: center bottom;
-    }
-
-  .List--ios::after {
-    top: 100%;
-    background: var(--separator_common);
-    transform-origin: center top;
-    }
-
-  .List--ios:empty::before,
-  .List--ios:empty::after {
-    content: none;
-    }
-
-@media (min-resolution: 2dppx) {
-  .List--ios::before,
-  .List--ios::after {
-    transform: scaleY(.5);
-    }
-  }
-
-@media (min-resolution: 3dppx) {
-  .List--ios::before,
-  .List--ios::after {
-    transform: scaleY(.33);
-    }
-  }
+.List {}
