+++ src/components/PanelHeaderContent/PanelHeaderContent.css
@@ -1,124 +1,99 @@
 .PanelHeaderContent {
   line-height: initial;
   height: 100%;
-  }
-
-  .PanelHeaderContent__in {
-    flex: 1;
-    min-width: 0;
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    height: 100%;
-    }
-
-  .PanelHeaderContent__status {
-    order: 1;
-    font-weight: 400;
-    }
-
-  .PanelHeaderContent__children {
-    max-width: 100%;
-    }
-
-    .PanelHeaderContent__children-in {
-      flex-grow: 0;
-      min-width: 0;
-      overflow: hidden;
-      white-space: nowrap;
-      text-overflow: ellipsis;
-      }
-
-/*
-  iOS
- */
-.PanelHeaderContent--ios {
-
-  }
-
-  .PanelHeaderContent--ios .PanelHeaderContent__in {
-    align-items: center;
-    justify-content: center;
-    }
-
-  .PanelHeaderContent--ios .PanelHeaderContent__in.Tappable--active {
-    opacity: .7;
-    }
-
-    .PanelHeaderContent--ios .PanelHeaderContent__children {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      font-size: 17px;
-      line-height: 20px;
-      }
-
-    .PanelHeaderContent--ios .PanelHeaderContent__status ~ .PanelHeaderContent__children {
-      font-size: 16px;
-      line-height: 19px;
-      }
-
-      .PanelHeaderContent--ios .PanelHeaderContent__aside {
-        margin-left: 2px;
-        display: flex;
-        align-items: center;
-        justify-content: flex-start;
-        }
-
-    .PanelHeaderContent--ios .PanelHeaderContent__status {
-      opacity: .6;
-      font-size: 13px;
-      }
-
-/*
-  Android
- */
-.PanelHeaderContent--android {
   max-width: 100%;
   display: flex;
   align-items: center;
   justify-content: flex-start;
   }
 
-  .PanelHeaderContent--android .PanelHeaderContent__before {
-    padding-right: 6px;
-    padding-left: 6px;
+  .PanelHeaderContent__before {
+    flex-shrink: 0;
+    margin-right: 12px;
     }
 
-  .PanelHeaderContent--android .PanelHeaderContent__in {
+  .PanelHeaderContent__in {
+    flex-grow: 1;
+    min-width: 0;
+    display: flex;
+    flex-direction: column;
     align-items: flex-start;
     justify-content: center;
-    padding: 0 6px;
+    height: 100%;
     }
+
+    .PanelHeaderContent__status {
+      order: 1;
+      font-weight: 400;
+      font-size: 13px;
+      line-height: 16px;
+      margin-top: 1px;
+      opacity: .6;
+      }
+
+    .PanelHeaderContent__status ~ .PanelHeaderContent__children .PanelHeaderContent__children-in {
+      font-weight: 500;
+      font-family: var(--font-common);
+      font-size: 16px;
+      line-height: 20px;
+      }
+
+    .PanelHeaderContent__children {
+      max-width: 100%;
+      display: flex;
+      align-items: center;
+      }
+
+      .PanelHeaderContent__children-in {
+        flex-grow: 0;
+        min-width: 0;
+        overflow: hidden;
+        white-space: nowrap;
+        text-overflow: ellipsis;
+        font-size: 21px;
+        line-height: var(--panelheader_height_ios);
+        font-weight: 800;
+        font-family: var(--font-tt);
+        }
+
+      .PanelHeaderContent__aside {
+        display: flex;
+        align-items: center;
+        justify-content: flex-start;
+        margin-left: 2px;
+        color: var(--header_tint);
+        }
+
+    .PanelHeaderContent__width {
+      width: 1000px;
+      }
+
+/*
+  iOS
+ */
+.PanelHeaderContent--ios {
+  font-family: var(--font-ios);
+  }
+
+  .PanelHeaderContent--ios .PanelHeaderContent__in.Tappable--active {
+    opacity: .7;
+    }
+
+  .PanelHeaderContent--ios .PanelHeaderContent__before ~ .PanelHeaderContent__in {
+    align-items: flex-start;
+    }
+
+  .PanelHeaderContent--ios .PanelHeaderContent__in {
+    align-items: center;
+    }
+
+/*
+  Android
+ */
+.PanelHeaderContent--android {
+  font-family: var(--font-android);
+  }
 
   .PanelHeaderContent--android .PanelHeaderContent__in.Tappable--active {
     background-color: rgba(255, 255, 255, .1);
     }
-
-    .PanelHeaderContent--android .PanelHeaderContent__children {
-      display: flex;
-      align-items: center;
-      justify-content: center;
-      font-size: 20px;
-      line-height: 28px;
-      font-weight: 500;
-      }
-
-    .PanelHeaderContent--android .PanelHeaderContent__status ~ .PanelHeaderContent__children {
-      font-size: 16px;
-      line-height: 19px;
-      }
-
-      .PanelHeaderContent--android .PanelHeaderContent__aside {
-        margin-left: 4px;
-        display: flex;
-        align-items: center;
-        justify-content: flex-start;
-        opacity: .8;
-        }
-
-    .PanelHeaderContent--android .PanelHeaderContent__status {
-      opacity: .6;
-      font-size: 13px;
-      line-height: 16px;
-      }
