+++ src/components/PanelHeaderContent/PanelHeaderContent.js
@@ -2,7 +2,6 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from '../../lib/classNames';
 import getClassName from '../../helpers/getClassName';
-import { IS_PLATFORM_ANDROID } from '../../lib/platform';
 import Tappable from '../Tappable/Tappable';
 
 const baseClassNames = getClassName('PanelHeaderContent');
@@ -14,13 +13,14 @@ const PanelHeaderContent = ({ className, style, aside, status, before, children,
 
   return (
     <div {...rootProps} className={classNames(baseClassNames, className)} style={style}>
-      {IS_PLATFORM_ANDROID && before && <div className="PanelHeaderContent__before">{before}</div>}
+      {before && <div className="PanelHeaderContent__before">{before}</div>}
       <InComponent {...inProps} className="PanelHeaderContent__in" onClick={onClick}>
         {status && <div className="PanelHeaderContent__status">{status}</div>}
         <div className="PanelHeaderContent__children">
           <span className="PanelHeaderContent__children-in">{children}</span>
           {aside && <div className="PanelHeaderContent__aside">{aside}</div>}
         </div>
+        <div className="PanelHeaderContent__width" />
       </InComponent>
     </div>
   );
@@ -33,9 +33,6 @@ PanelHeaderContent.propTypes = {
   aside: PropTypes.node,
   status: PropTypes.node,
   onClick: PropTypes.func,
-  /**
-   * Android only
-   */
   before: PropTypes.node
 };
