+++ src/components/PanelHeaderContent/Readme.md
@@ -9,9 +9,7 @@
     >
       <PanelHeaderContent
         status="был в сети сегодня, в 18:46"
-        aside={<Icon16Dropdown />}
-        before={<Avatar size={40} src="https://sun9-5.userapi.com/c834100/v834100961/4f8f1/hjsBzq433co.jpg?ava=1" />}
-        onClick={() => {}}
+        before={<Avatar size={36} src="https://sun9-5.userapi.com/c834100/v834100961/4f8f1/hjsBzq433co.jpg?ava=1" />}
       >
         Влад Анесов
       </PanelHeaderContent>
