+++ src/components/Search/Readme.md
@@ -1,10 +1,3 @@
-Существует два способа разместить поиск – в теле `Panel` и в шапке `PanelHeader`. Расположение поиска влияет на
-его внешний вид. Для правильной отрисовки ему нужно указать свойство `theme`.
-
-**Важно:** нельзя размещать `<Search theme="header" />` внутри панели. Нельзя размещать `<Search theme="default" />`
-в шапке.
-
-
 ```jsx
   const thematics = [
     {id: 3201, name: "Аренда автомобилей"},
@@ -72,20 +65,19 @@
 
     render() {
       return (
-        <div>
+        <React.Fragment>
           <PanelHeader
-            noShadow
             right={<HeaderButton onClick={this.props.goHeaderSearch} key="add"><Icon24Add /></HeaderButton>}
           >
             Выбор тематики
           </PanelHeader>
-          <Search value={this.state.search} onChange={this.onChange}/>
+          <Search value={this.state.search} onChange={this.onChange} after="Отмена"/>
           {this.thematics.length > 0 &&
             <List>
               {this.thematics.map(thematic => <Cell key={thematic.id}>{thematic.name}</Cell>)}
             </List>
           }
-        </div>
+        </React.Fragment>
       );
     }
   }
@@ -95,15 +87,11 @@
     constructor (props) {
       super(props);
       this.state = {
-        showSearch: osname === IOS,
         search: ''
       }
-      this.toggleSearch = this.toggleSearch.bind(this);
 
       this.onChange = this.onChange.bind(this);
     }
-
-    toggleSearch () { this.setState({ showSearch: !this.state.showSearch }); }
 
     onChange (search) { this.setState({ search }); }
 
@@ -114,19 +102,15 @@
 
     render () {
       return (
-        <div>
+        <React.Fragment>
           <PanelHeader
-            left={<HeaderButton onClick={this.props.goSearch}>{osname === IOS ? <Icon28ChevronBack /> : <Icon24Back />}</HeaderButton>}
-            right={osname === ANDROID && <HeaderButton onClick={this.toggleSearch}><Icon24Search /></HeaderButton>}
+            left={<PanelHeaderBack onClick={this.props.goSearch} />}
           >
-            {this.state.showSearch ?
-              <Search
-                theme="header"
-                value={this.state.search}
-                onChange={this.onChange}
-                onClose={this.toggleSearch}
-              /> : 'Поиск'
-            }
+            <Search
+              theme="header"
+              value={this.state.search}
+              onChange={this.onChange}
+            />
           </PanelHeader>
           <List>
             {this.users.map((user) => (
@@ -137,7 +121,7 @@
               >{user.name}</Cell>
             ))}
           </List>
-        </div>
+        </React.Fragment>
       );
     }
   }
@@ -161,10 +145,10 @@
     render () {
       return (
         <View activePanel={this.state.activePanel}>
-          <Panel id="search">
+          <Panel id="search" separator={false}>
             <SimpleSearch goHeaderSearch={this.goHeaderSearch}/>
           </Panel>
-          <Panel id="header-search">
+          <Panel id="header-search" separator={false}>
             <HeaderSearch goSearch={this.goSearch}/>
           </Panel>
         </View>
