+++ src/components/Search/Search.css
@@ -1,0 +1,225 @@
+.Search {
+  overflow: hidden;
+  font-weight: normal;
+  font-family: var(--font-common);
+  transform: translate3d(0, 0, 0);
+  line-height: normal;
+  }
+
+  .Search__in {
+    position: relative;
+    }
+
+    .Search__width {
+      height: var(--search_default_height);
+      white-space: nowrap;
+      color: transparent;
+      width: 10000px;
+      }
+
+    .Search__control {
+      display: flex;
+      justify-content: flex-start;
+      align-items: center;
+      position: absolute;
+      left: 0;
+      top: 0;
+      width: calc(100% - 1px);
+      height: 100%;
+      border-radius: 10px;
+      }
+
+      .Search__input {
+        border: none;
+        margin: 0;
+        height: var(--search_default_height);
+        position: relative;
+        padding: 0 22px 0 36px;
+        box-sizing: border-box;
+        font-size: 17px;
+        outline: none;
+        border-radius: 10px;
+        max-width: 100%;
+        flex-grow: 1;
+        min-width: 0;
+        background-color: transparent;
+        z-index: 2;
+        color: var(--text_primary);
+        }
+
+      .Search--has-after .Search__input {
+        border-radius: 10px 0 0 10px;
+        padding-right: 10px;
+        }
+
+      .Search__after-width {
+        font-size: 17px;
+        height: var(--search_default_height);
+        line-height: var(--search_default_height);
+        padding-left: 12px;
+        padding-right: 4px;
+        }
+
+      .Search__placeholder {
+        position: absolute;
+        left: 0;
+        width: 100%;
+        cursor: text;
+        z-index: 1;
+        text-align: left;
+        display: block;
+        height: var(--search_default_height);
+        font-size: 0;
+        padding-left: 12px;
+        }
+
+        .Search__placeholder-in {
+          height: var(--search_default_height);
+          display: flex;
+          align-items: center;
+          }
+
+          .Search__placeholder-text {
+            margin-left: 8px;
+            line-height: 22px;
+            font-size: 17px;
+            }
+
+          .Search--has-value .Search__placeholder-text {
+            opacity: 0;
+            }
+
+    .Search__after {
+      position: absolute;
+      left: 100%;
+      top: 0;
+      line-height: var(--search_default_height);
+      z-index: 2;
+      transform: translate(0);
+      transition: transform .3s var(--ios-easing);
+      cursor: default;
+      }
+
+    .Search--focused .Search__after,
+    .Search--has-value .Search__after {
+      transform: translate(-100%);
+      }
+
+      .Search__after::before {
+        position: absolute;
+        right: calc(100% - 1px);
+        display: block;
+        height: 100%;
+        width: 10px;
+        content: '';
+        pointer-events: none;
+        }
+
+      .Search__after-in {
+        position: relative;
+        z-index: 2;
+        cursor: pointer;
+        font-size: 17px;
+        padding-left: 12px;
+        padding-right: 4px;
+        }
+
+      .Search__after::after {
+        position: absolute;
+        right: 100%;
+        display: block;
+        height: 100%;
+        width: 10px;
+        content: '';
+        top: 0;
+        border-radius: 0 10px 10px 0;
+        cursor: text;
+        pointer-events: none;
+        }
+
+/*
+  Themes
+ */
+
+/*
+  default
+ */
+.Search {
+  background: var(--search_bar_background);
+  padding: 8px;
+  }
+
+  .Search__control {
+    background-color: var(--search_bar_field_background);
+    }
+
+    .Search__input {
+      color: var(--text_primary);
+      }
+
+    .Search__placeholder {
+      color: var(--search_bar_field_tint);
+      }
+
+    .Search__after-width {
+      background: var(--search_bar_field_background);
+      color: var(--search_bar_field_background);
+      }
+
+  .Search__after {
+    background: var(--search_bar_background);
+    color: var(--accent);
+    }
+
+    .Search__after::after {
+      background-color: var(--search_bar_field_background);
+      }
+
+    .Search__after::before {
+      background-color: var(--search_bar_background);
+      }
+
+/*
+  header
+ */
+.PanelHeader .Search {
+  background: var(--header_background);
+  padding: 0;
+  }
+
+  .PanelHeader .Search__control {
+    background-color: var(--header_search_field_background);
+    }
+
+    .PanelHeader .Search__input {
+      color: var(--text_primary);
+      }
+
+    .PanelHeader .Search__placeholder {
+      color: var(--header_search_field_tint);
+      }
+
+    .PanelHeader .Search__after-width {
+      background: var(--header_search_field_background);
+      color: var(--header_search_field_background);
+      }
+
+  .PanelHeader .Search__after {
+    background: var(--header_background);
+    color: var(--header_tint);
+    }
+
+    .PanelHeader .Search__after::after {
+      background-color: var(--header_search_field_background);
+      }
+
+    .PanelHeader .Search__after::before {
+      background-color: var(--header_background);
+      }
+
+/**
+  Android
+ */
+.Search--android.Search--default {
+  padding: 8px 16px;
+  }
