--- src/components/SearchAndroid/SearchAndroid.css
@@ -1,104 +1,0 @@
-.Search--android {
-  height: 56px;
-  box-sizing: border-box;
-  width: 100%;
-  background: var(--header_background);
-  }
-
-  .Search--android .Search__container {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    }
-
-  .Search--android .Search__before,
-  .Search--android .Search__after {
-    flex-shrink: 0;
-    }
-
-  .Search--android .Search__control {
-    max-width: 100%;
-    flex-grow: 1;
-    }
-
-    .Search--android .Search__input {
-      border: none;
-      display: block;
-      width: 100%;
-      padding: 0;
-      box-sizing: border-box;
-      outline: none;
-      font-size: 16px;
-      line-height: 20px;
-      background-color: transparent;
-      }
-
-/*
-  header theme
- */
-
-.Search--android.Search--header {
-  padding: 4px;
-  }
-
-.Search--android.Search--header .Search__before {
-  margin-right: 12px;
-  }
-
-.Search--android.Search--header .Search__after {
-  margin-left: 12px;
-  }
-
-.Search--android.Search--header.Search--vkapps .Search__after {
-  margin-right: 90px;
-  }
-
-.Search--android.Search--header .Search__input {
-  color: var(--header_tint);
-  padding: 15px 0 13px;
-  }
-
-.Search--android.Search--header .Search__input::placeholder {
-  color: var(--header_search_field_tint);
-  opacity: .6;
-  }
-
-/*
-  default theme
- */
-
-.Search--android.Search--default {
-  padding: 8px;
-  box-shadow: 0 0 4px rgba(0, 0, 0, .08), 0 4px 4px rgba(0, 0, 0, .16);
-  position: relative;
-  z-index: 2;
-  }
-
-.Search--android.Search--default .Search__container {
-  background-color: var(--header_search_field_background);
-  border-radius: 8px;
-  padding: 0 4px;
-  }
-
-.Search--android.Search--default .Search__before {
-  margin-right: 8px;
-  }
-
-.Search--android.Search--default .Search__after {
-  margin-left: 8px;
-  }
-
-.Search--android.Search--default .Search__before .Icon,
-.Search--android.Search--default .Search__after .Icon {
-  padding: 8px;
-  color: var(--icon_secondary);
-  }
-
-.Search--android.Search--default .Search__input {
-  padding: 11px 0 9px;
-  color: var(--header_tint);
-  }
-
-.Search--android.Search--default .Search__input::placeholder {
-  color: var(--icon_secondary);
-  }
