--- src/components/SearchIOS/SearchIOS.js
@@ -1,152 +1,0 @@
-import React from 'react';
-import PropTypes from 'prop-types';
-import getClassName from '../../helpers/getClassName';
-import classNames from '../../lib/classNames';
-import Icon16SearchOutline from '@vkontakte/icons/dist/16/search_outline';
-
-const baseClassName = getClassName('Search');
-
-let searchId = 0;
-
-export default class SearchIOS extends React.Component {
-  constructor (props) {
-    super(props);
-
-    let state = {
-      showAfter: false,
-      focused: false,
-      placeholderOffset: null,
-      afterWidth: null
-    };
-
-    if (props.hasOwnProperty('value')) {
-      this.isControlledOutside = true;
-    } else {
-      state.value = props.defaultValue || '';
-    }
-    searchId++;
-    this.state = state;
-  }
-
-  static propTypes = {
-    className: PropTypes.string,
-    after: PropTypes.node,
-    defaultValue: PropTypes.string,
-    value: PropTypes.string,
-    onChange: PropTypes.func,
-    onFocus: PropTypes.func,
-    onBlur: PropTypes.func,
-    placeholder: PropTypes.node,
-    theme: PropTypes.oneOf(['header', 'default']),
-    getRef: PropTypes.oneOfType([
-      PropTypes.func,
-      PropTypes.shape({ current: PropTypes.any })
-    ])
-  };
-
-  static defaultProps = {
-    after: 'Отменить',
-    theme: 'default',
-    placeholder: 'Поиск'
-  };
-
-  get value () {
-    return this.isControlledOutside ? this.props.value : this.state.value;
-  }
-
-  onFocus = (e) => {
-    this.setState({ focused: true });
-    this.props.onFocus && this.props.onFocus(e);
-  };
-
-  onBlur = (e) => {
-    this.setState({ focused: false });
-    this.props.onBlur && this.props.onBlur(e);
-  };
-
-  onChange = (e) => {
-    if (!this.isControlledOutside) {
-      this.setState({ value: e.target.value });
-    }
-    if (this.props.onChange) {
-      this.props.onChange(e.target.value, e);
-    }
-  };
-
-  onCancel = () => {
-    if (!this.isControlledOutside) {
-      this.setState({ value: '' });
-    }
-    if (this.props.onChange) {
-      this.props.onChange('');
-    }
-  };
-
-  inputRef = (element) => {
-    const getRef = this.props.getRef;
-    if (getRef) {
-      if (typeof getRef === 'function') {
-        getRef(element);
-      } else {
-        getRef.current = element;
-      }
-    }
-  };
-
-  render () {
-    const {
-      className,
-      theme,
-      onFocus,
-      onBlur,
-      onChange,
-      defaultValue,
-      value,
-      placeholder,
-      after,
-      getRef,
-      ...inputProps
-    } = this.props;
-
-    return (
-      <div className={classNames(baseClassName, {
-        [`Search--${theme}`]: true,
-        'Search--focused': this.state.focused,
-        'Search--has-value': this.value,
-        'Search--has-after': after
-      }, className)}>
-        <div className="Search__in">
-          <div className="Search__width" />
-          <div className="Search__control">
-            <input
-              {...inputProps}
-              id={`search-${searchId}`}
-              ref={this.inputRef}
-              type="text"
-              className="Search__input"
-              onFocus={this.onFocus}
-              onBlur={this.onBlur}
-              onChange={this.onChange}
-              value={this.value}
-            />
-            {after && <div className="Search__after-width">{after}</div>}
-            <label
-              className="Search__placeholder"
-              htmlFor={`search-${searchId}`}
-            >
-              <div className="Search__placeholder-in">
-                <Icon16SearchOutline />
-                <div className="Search__placeholder-text" dangerouslySetInnerHTML={{ __html: placeholder }} />
-              </div>
-            </label>
-          </div>
-          {after &&
-          <div className="Search__after" onClick={this.onCancel}>
-            <div className="Search__after-in">{after}</div>
-          </div>
-          }
-        </div>
-      </div>
-    );
-  }
-}
