+++ src/components/Snackbar/Readme.md
@@ -97,7 +97,7 @@ class SnackBarExample extends React.Component {
   render() {
     return (
       <View header={false} activePanel="example">
-        <Panel id="example">
+        <Panel id="example" separator={false}>
           <Group>
             <CellButton onClick={this.openBase}>Простое уведомление с иконкой</CellButton>
             <CellButton onClick={this.openBaseWithAction}>Уведомление с иконкой и кнопкой</CellButton>
