+++ src/components/Switch/Readme.md
@@ -1,21 +1,19 @@
-```
+```jsx harmony
   <View activePanel="switch">
     <Panel id="switch">
       <PanelHeader>
         Switch
       </PanelHeader>
       <Group>
-        <List>
-          <Cell asideContent={<Switch />}>
-            Комментарии к записям
-          </Cell>
-          <Cell asideContent={<Switch defaultChecked />}>
-            Ссылки
-          </Cell>
-          <Cell asideContent={<Switch disabled />}>
-            Фотоальбомы
-          </Cell>
-        </List>
+        <Cell asideContent={<Switch />}>
+          Комментарии к записям
+        </Cell>
+        <Cell asideContent={<Switch defaultChecked />}>
+          Ссылки
+        </Cell>
+        <Cell asideContent={<Switch disabled />}>
+          Фотоальбомы
+        </Cell>
       </Group>
     </Panel>
   </View>
