+++ src/components/Tabs/Tabs.css
@@ -1,11 +1,24 @@
+.Tabs {
+  background: var(--background_content);
+  }
+
 .Tabs,
 .Tabs .HorizontalScroll__in {
   display: flex;
   align-items: center;
   }
 
+.PanelHeader .Tabs {
+  background: var(--header_background);
+  }
+
 .Tabs .HorizontalScroll {
   min-width: 100%;
+  }
+
+.Tabs--buttons {
+  padding-top: 8px;
+  padding-bottom: 8px;
   }
 
 .Tabs--default,
@@ -18,72 +31,25 @@
   justify-content: flex-start;
   }
 
+.Tabs.Tabs--buttons .HorizontalScroll__in::after,
+.Tabs.Tabs--buttons .HorizontalScroll__in::before {
+  display: block;
+  width: 8px;
+  flex-shrink: 0;
+  height: 1px;
+  content: '';
+  }
+
 /*
  iOS
  */
 
-.Tabs--ios.Tabs--default.Tabs--light {
-  padding: 8px;
-  background: var(--search_bar_background);
+.Tabs--ios {
+  font-family: var(--font-ios);
   }
 
-.Tabs--ios.Tabs--default.Tabs--header {
-  padding-top: 6px;
-  }
-
-.Tabs--ios.Tabs--buttons.Tabs--header {
-  background: var(--header_background);
-  }
-
-.Tabs--ios.Tabs--buttons {
-  padding-top: 8px;
-  padding-bottom: 8px;
-  }
-
-.FixedLayout--ios.FixedLayout--top .Tabs::before {
-  position: absolute;
-  right: 0;
-  left: 0;
-  z-index: 1;
-  height: 1px;
-  content: '';
-  bottom: 0;
-  background: var(--separator_common);
-  transform-origin: center bottom;
-  }
-
-.FixedLayout--ios.FixedLayout--bottom .Tabs::before {
-  position: absolute;
-  right: 0;
-  left: 0;
-  z-index: 1;
-  height: 1px;
-  content: '';
-  bottom: 100%;
-  background: var(--separator_common);
-  transform-origin: center bottom;
-  }
-
-@media (min-resolution: 2dppx) {
-  .FixedLayout--ios.FixedLayout--top .Tabs::before,
-  .FixedLayout--ios.FixedLayout--bottom .Tabs::before {
-    transform: scaleY(.5);
-    }
-  }
-
-@media (min-resolution: 3dppx) {
-  .FixedLayout--ios.FixedLayout--top .Tabs::before,
-  .FixedLayout--ios.FixedLayout--bottom .Tabs::before {
-    transform: scaleY(.33);
-    }
-  }
-
-.FixedLayout--ios.FixedLayout--top .Tabs--light.Tabs--buttons {
-  background: var(--background_content);
-  }
-
-.FixedLayout--ios.FixedLayout--bottom .Tabs {
-  background: var(--header_alternate_background);
+.Tabs--ios.Tabs--segmented {
+  padding: 4px 12px 8px;
   }
 
 /* Нужно для закрашивания прозрачной области нижнего инсета на iphone x */
@@ -94,100 +60,51 @@
   width: 100%;
   height: 100%;
   content: '';
-  background-color: var(--header_alternate_background);
+  background: var(--background_content);
   }
 
-  .FixedLayout--ios.FixedLayout--bottom .TabsItem {
+  .FixedLayout--ios.FixedLayout--bottom .TabsItem,
+  .FixedLayout--ios.FixedLayout--bottom .Separator {
     position: relative;
     z-index: 2;
     }
-
-.Tabs--ios.Tabs--buttons .HorizontalScroll__in::after,
-.Tabs--ios.Tabs--buttons .HorizontalScroll__in::before {
-  display: block;
-  width: 8px;
-  flex-shrink: 0;
-  height: 1px;
-  content: '';
-  }
 
 /*
   Android
  */
 .Tabs--android {
-
-  }
-
-.Tabs--android.Tabs--light {
-  background-color: var(--background_content);
-  }
-
-.Tabs--android.Tabs--header {
-  background-color: var(--header_background);
-  }
-
-.Tabs--android.Tabs--buttons .HorizontalScroll__in::after,
-.Tabs--android.Tabs--buttons .HorizontalScroll__in::before {
-  display: block;
-  width: 8px;
-  flex-shrink: 0;
-  height: 1px;
-  content: '';
-  }
-
-.FixedLayout--top .Tabs--android,
-.FixedLayout--bottom .Tabs--android {
-  box-shadow: 0 0 4px rgba(0, 0, 0, .08), 0 4px 4px rgba(0, 0, 0, .16);
-  }
-
-.FixedLayout--android.FixedLayout--bottom .Tabs {
-  background: var(--header_alternate_background);
+  font-family: var(--font-android);
   }
 
 /**
  * Counter inside TabsItem
  */
-.Tabs--android.Tabs--light .TabsItem .Counter {
-  background: var(--counter_secondary_background);
-  color: var(--counter_secondary_text);
+
+.TabsItem .Counter {
+  margin-left: 8px;
   }
 
-.Tabs--android.Tabs--light .TabsItem--selected .Counter {
-  background: var(--counter_primary_background);
-  color: var(--counter_primary_text);
-  }
-
-.Tabs--ios.Tabs--light .TabsItem .Counter {
-  background: var(--segmented_control_tint);
-  color: var(--background_content);
-  }
-
-.Tabs--ios.Tabs--light .TabsItem--selected .Counter {
-  background: var(--background_content);
-  color: var(--segmented_control_tint);
-  }
-
-.Tabs--ios.Tabs--buttons.Tabs--light .TabsItem .Counter {
+.Tabs .TabsItem .Counter {
   background: var(--panel_tab_inactive_text);
   color: var(--white);
   }
 
-.Tabs--ios.Tabs--buttons.Tabs--light .TabsItem--selected .Counter {
+.Tabs .TabsItem--selected .Counter {
   background: var(--panel_tab_active_text);
   color: var(--panel_tab_active_background);
   }
 
-.Tabs--buttons.Tabs--header .TabsItem .Counter {
-  background: var(--header_tab_inactive_text);
-  color: var(--header_background);
+.Tabs--ios.Tabs--segmented .TabsItem .Counter {
+  background: var(--segmented_control_tint);
+  color: var(--background_content);
   }
 
-.Tabs--buttons.Tabs--header .TabsItem .Counter {
-  background: var(--header_tab_inactive_text);
-  color: var(--header_background);
+.Tabs--ios.Tabs--segmented .TabsItem--selected .Counter {
+  background: var(--background_content);
+  color: var(--segmented_control_tint);
   }
 
-.Tabs--buttons.Tabs--header .TabsItem--selected .Counter {
-  background: var(--header_tab_active_text);
-  color: var(--header_background);
+.Tabs--ios.Tabs--segmented .TabsItem .Counter--s-m {
+  position: relative;
+  top: 1px;
   }
