+++ src/components/Tabs/Tabs.js
@@ -10,13 +10,9 @@ export default class Tabs extends React.Component {
     children: PropTypes.node,
     className: PropTypes.string,
     /**
-     * Значения white и gray устарели. Они будут удалены в следующей мажорной версии
+     * тип segmented доступен только для iOS
      */
-    theme: PropTypes.oneOf(['light', 'header']),
-    /**
-     * Свойство отвечает за растягивание табов
-     */
-    type: PropTypes.oneOf(['default', 'buttons']),
+    type: PropTypes.oneOf(['default', 'buttons', 'segmented']),
     style: PropTypes.object,
     getRootRef: PropTypes.oneOfType([
       PropTypes.func,
@@ -25,33 +21,19 @@ export default class Tabs extends React.Component {
   };
 
   static defaultProps = {
-    theme: 'light',
     type: 'default'
   };
 
-  /**
-   * Метод нужен для маппинга устаревших значений theme на новые
-   * @returns {string}
-   */
-  get theme () {
-    switch (this.props.theme) {
-      case 'white':
-        return 'header';
-      case 'gray':
-        return 'light';
-      default:
-        return this.props.theme;
-    }
-  }
-
   render () {
-    const { className, children, theme, style, type, getRootRef, ...restProps } = this.props;
+    const { className, children, style, type, getRootRef, ...restProps } = this.props;
 
     return (
-      <div {...restProps} ref={getRootRef} className={classNames(baseClassName, {
-        [`Tabs--${this.theme}`]: true,
-        [`Tabs--${type}`]: true
-      }, className)} style={style}>{children}</div>
+      <div
+        {...restProps}
+        ref={getRootRef}
+        className={classNames(baseClassName, `Tabs--${type}`, className)}
+        style={style}
+      >{children}</div>
     );
   }
 }
