--- src/lib/webview.ts
@@ -1,3 +1,0 @@
-import vkConnect from '@vkontakte/vk-connect';
-
-export const isWebView: boolean = vkConnect.isWebView();
