+++ src/styles/constants.css
@@ -3,6 +3,7 @@
   --font-ios: -apple-system, Helvetica Neue, Arial;
   --font-android: Roboto, Arial;
   --font-common: -apple-system, Roboto, Helvetica Neue, Arial;
+  --font-tt: "TT Commons", -apple-system, Roboto, Helvetica Neue, Arial;
 
   /* easings */
   --ios-easing: cubic-bezier(.36, .66, .04, 1);
@@ -10,11 +11,11 @@
 
   /* sizes */
   --tabbar_height: 48px;
-  --panelheader_height_ios: 44px;
+  --panelheader_height_ios: 52px;
   --panelheader_height_android: 56px;
   --modalheader_height_ios: 52px;
   --modalheader_height_android: 56px;
-  --search_default_height_ios: 32px;
+  --search_default_height: 36px;
 
   /* colors */
   --white: #fff;
