+++ src/styles/styles.css
@@ -53,8 +53,7 @@
 @import '../components/Avatar/Avatar.css';
 @import '../components/Gallery/Gallery.css';
 @import '../components/Progress/Progress.css';
-@import '../components/SearchIOS/SearchIOS.css';
-@import '../components/SearchAndroid/SearchAndroid.css';
+@import '../components/Search/Search.css';
 @import '../components/Tabs/Tabs.css';
 @import '../components/TabsItem/TabsItem.css';
 @import '../components/Tooltip/Tooltip.css';
@@ -76,6 +75,12 @@
 @import '../components/Select/Select.css';
 @import '../components/Textarea/Textarea.css';
 @import '../components/File/File.css';
+
+@font-face {
+  font-family: "TT Commons";
+  src: url("../fonts/TTCommons-DemiBold.woff") format("woff");
+  font-weight: 800;
+  }
 
 html,
 body,
